+++ CHANGELOG.md
@@ -7,8 +7,15 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [v0.8.0] - 2022-08-30
+## [v0.8.1] - 2022-11-23
+### Fixed
+- [#26] Use correct label for dogu resources
+  - `dogu.name=name` is now valid
 
+### Changed
+- [#24] Read dogu registry URL schema from cluster secret instead of config.
+
+## [v0.8.0] - 2022-08-30
 ### Changed
 - [#22] If the resource urls from the k8s-components e.g. `dogu-operator` have the same host as the configured
   dogu registry, basic auth will be used for those components.